:recycle: request/response 위치 이동 #LF1-432
