+++ src/main/java/kr/bb/store/domain/store/repository/StoreRepositoryCustom.java
@@ -3,7 +3,7 @@ package kr.bb.store.domain.store.repository;
 import kr.bb.store.domain.store.entity.Store;
 import kr.bb.store.domain.store.entity.address.Gugun;
 import kr.bb.store.domain.store.entity.address.Sido;
-import kr.bb.store.domain.store.handler.response.StoreForMapResponse;
+import kr.bb.store.domain.store.controller.response.StoreForMapResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
