+++ src/main/java/kr/bb/store/domain/store/repository/StoreRepositoryCustomImpl.java
@@ -6,7 +6,7 @@ import kr.bb.store.domain.store.entity.Store;
 import kr.bb.store.domain.store.entity.address.Gugun;
 import kr.bb.store.domain.store.entity.address.Sido;
 import kr.bb.store.domain.store.handler.response.QStoreForMapResponse;
-import kr.bb.store.domain.store.handler.response.StoreForMapResponse;
+import kr.bb.store.domain.store.controller.response.StoreForMapResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
