+++ src/test/java/kr/bb/store/domain/store/handler/StoreCreatorTest.java
@@ -2,7 +2,7 @@ package kr.bb.store.domain.store.handler;
 
 import kr.bb.store.domain.store.entity.Store;
 import kr.bb.store.domain.store.exception.CannotOwnMultipleStoreException;
-import kr.bb.store.domain.store.handler.request.StoreRequest;
+import kr.bb.store.domain.store.dto.StoreRequest;
 import kr.bb.store.domain.store.repository.StoreRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
