+++ src/test/java/kr/bb/store/domain/store/handler/StoreReaderTest.java
@@ -8,10 +8,10 @@ import kr.bb.store.domain.store.entity.address.GugunRepository;
 import kr.bb.store.domain.store.entity.address.Sido;
 import kr.bb.store.domain.store.entity.address.SidoRepository;
 import kr.bb.store.domain.store.exception.StoreNotFoundException;
-import kr.bb.store.domain.store.handler.response.StoreDetailInfoResponse;
-import kr.bb.store.domain.store.handler.response.StoreInfoManagerResponse;
-import kr.bb.store.domain.store.handler.response.StoreInfoUserResponse;
-import kr.bb.store.domain.store.handler.response.StoreListForMapResponse;
+import kr.bb.store.domain.store.controller.response.StoreDetailInfoResponse;
+import kr.bb.store.domain.store.controller.response.StoreInfoManagerResponse;
+import kr.bb.store.domain.store.controller.response.StoreInfoUserResponse;
+import kr.bb.store.domain.store.controller.response.StoreListForMapResponse;
 import kr.bb.store.domain.store.repository.DeliveryPolicyRepository;
 import kr.bb.store.domain.store.repository.StoreAddressRepository;
 import kr.bb.store.domain.store.repository.StoreRepository;
